Better design: buffer_t is now a shared_ptr.
The buffers were being deleted formerly, but now they are preserved
properly by being converted to std::shared_ptr. This way the vector
doesn't make copies that lose the state (since stuff is being stored on
the drive).

Now the packages seem to be created correctly and read correctly, while
being swapped to the drive instead of managed in memory. I have it
hardcoded for now to always go to the tmp folder, but I will add next
support for mimimum ram storage and swap to file for larger packages (>=
1GB).

There is still a bug--the files in /tmp/wpkg are not being cleaned up
properly--this means the destructors are not being called (which leads
me to believe there could be a memory leak in the management of buffers,
but I cannot see how right now).

Also, wpkg_filename had thrown a memfile exceptions for IO errors, when
I believe the author intended to throw a wpkg_filename_exception_io
object instead. Seems like a typo to me, so it's fixed.