Improved swapping algorithm, and added mem percentage output.
Now swaps out buffers on write, only after a certain time has passed
since the last read. Reading buffers back in removes the buffer file
from the disk, and resets the count.

Added output to show the current system memory and process usage. This
is a work in progress which will allow me to determine the process size
and the optimim time to swap out old buffers to disk.