Improve create_index speed by using the pre-generated md5sum file
Generate md5sum file when generating the pkg and deploy in the repo.

The goal is to have a much faster create_index call because in this case
the md5sum are not re-generated each time. It compares the file with
the md5 provided by the package. If nothing change, the index entry is
not recomputed.